gadgets: Improve UX of adding a gadget.
Bring UX of adding a gadget close to ig-desktop.
