+++ src/gadgets/gadgetcreationinresource.tsx
@@ -3,6 +3,7 @@ import { Box, IconButton, Modal, Paper, Typography } from '@mui/material';
 import React from 'react';
 import { fetchInspektorGadgetFromArtifactHub } from '../api/artifacthub';
 import { GadgetCardEmbedWrapper, GadgetGrid } from './gadgetGrid';
+import GadgetInput from './gadgetInput';
 
 export function GadgetCreation({ resource, open, setOpen }) {
   const [gadgets, setGadgets] = React.useState([]);
@@ -32,26 +33,26 @@ export function GadgetCreation({ resource, open, setOpen }) {
         >
           <Box sx={{ display: 'flex', justifyContent: 'flex-end', alignItems: 'center', mb: 1 }}>
             <Typography variant="h6" sx={{ flexGrow: 1 }}>
-              Add Gadget
+              Gadget Gallery
             </Typography>
             <IconButton onClick={() => setOpen(false)} size="small">
               <Icon icon="mdi:close" />
             </IconButton>
           </Box>
-
+            <Box sx={{ mb:8, display: 'flex', flexDirection: 'column', gap: 1 }}>
+              <Typography variant="caption" sx={{ fontSize: '1rem' }}>
+                Enter a gadget image URL or discover gadgets from ArtifactHub
+              </Typography>
+              <GadgetInput resource={''} onAddGadget={() => {}} />
+            </Box>
           <Box sx={{ overflow: 'auto', flexGrow: 1 }}>
             <GadgetGrid
               resource={resource}
               gadgets={gadgets}
-              enableEmbed={false}
-              onViewSelect={() => {
-                // setOpen(true);
-              }}
               onEmbedClick={gadget => {
                 setSelectedGadget(gadget);
                 setEmbedDialogOpen(true);
               }}
-              callbackRunGadget={() => {}}
             />
             {embedDialogOpen && (
               <GadgetCardEmbedWrapper
